requirements.txt
streamlit
flask==2.3.3
python-docx==0.8.11
pandas==2.0.3
requests==2.31.0
openpyxl
gunicorn==21.2.0